Rename dancer view model file and update imports
